+++ src/sentry/performance_issues/detectors/io_main_thread_detector.py
@@ -18,9 +18,8 @@ from sentry.models.debugfile import ProjectDebugFile
 from sentry.models.organization import Organization
 from sentry.models.project import Project
 
-from ..base import (
-    DetectorType,
-    PerformanceDetector,
+from ..base import DetectorType, PerformanceDetector
+from ..detectors.utils import (
     get_notification_attachment_body,
     get_span_evidence_value,
     total_span_time,
