+++ src/sentry/performance_issues/detectors/mn_plus_one_db_span_detector.py
@@ -15,9 +15,8 @@ from sentry.issues.issue_occurrence import IssueEvidence
 from sentry.models.organization import Organization
 from sentry.models.project import Project
 
-from ..base import (
-    DetectorType,
-    PerformanceDetector,
+from ..base import DetectorType, PerformanceDetector
+from ..detectors.utils import (
     get_notification_attachment_body,
     get_span_evidence_value,
     total_span_time,
