+++ src/sentry/performance_issues/detectors/uncompressed_asset_detector.py
@@ -9,9 +9,8 @@ from sentry.issues.issue_occurrence import IssueEvidence
 from sentry.models.organization import Organization
 from sentry.models.project import Project
 
-from ..base import (
-    DetectorType,
-    PerformanceDetector,
+from ..base import DetectorType, PerformanceDetector
+from ..detectors.utils import (
     fingerprint_resource_span,
     get_notification_attachment_body,
     get_span_duration,
